backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "deepgram-sdk>=5.0.0",
    "fastapi>=0.100.0",
    "groq>=1.0.0",
    "httpx>=0.25.0",
    "loguru>=0.7.2",
    "numpy>=1.24.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.3.0",
    "pydub>=0.25.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "sounddevice>=0.4.6",
    "typing-extensions>=4.9.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
]
